Update README.md
The project uses a mechanical arm composed of two rotating joints to achieve movement along a trajectory. Users can generate movement trajectories in two ways: one is to generate a specified trajectory with a single key press, and the other is to draw a custom trajectory on the screen. The environment in which the project runs is as follows: Platform: Windows 11, Python 3.13.7, numpy 2.3.4, matplotlib 3.10.7. The following is the demonstration of the project: [images].